Refactor update user forms
